+++ pkg/deployment/agency/cache_test.go
@@ -17,36 +17,5 @@
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
 //
-// Author Adam Janikowski
-//
 
 package agency
-
-import (
-	"fmt"
-	"strings"
-)
-
-const (
-	ArangoKey = "arango"
-
-	PlanKey    = "Plan"
-	CurrentKey = "Current"
-
-	PlanCollectionsKey = "Collections"
-
-	SupervisionKey            = "Supervision"
-	SupervisionMaintenanceKey = "Maintenance"
-)
-
-func GetAgencyKey(parts ...string) string {
-	return fmt.Sprintf("/%s", strings.Join(parts, "/"))
-}
-
-func GetAgencyReadKey(elements ...string) []string {
-	return elements
-}
-
-func GetAgencyReadRequest(elements ...[]string) [][]string {
-	return elements
-}